requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.29.0
fonttools==4.62.1
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.47
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.12.0
idna==3.13
Jinja2==3.1.6
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.1
pillow==12.2.0
platformdirs==4.9.6
protobuf==7.34.1
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
scipy==1.17.1
segmentation_models_pytorch==0.5.0
sentry-sdk==2.58.0
setuptools==81.0.0
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
smmap==5.0.3
stringzilla==4.6.0
sympy==1.14.0
timm==1.0.26
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wandb==0.26.1
